Rename packages to 'task' and 'command'
