+++ src/main/java/duke/command/Bye.java
@@ -1,4 +1,4 @@
-package duke.commands;
+package duke.command;
 
 import duke.*;
