+++ src/main/java/duke/command/List.java
@@ -1,4 +1,4 @@
-package duke.commands;
+package duke.command;
 
 import duke.*;
